move ActionButtonWithTooltip to Dashboard/common to fix layering
