+++ src/components/Dashboard/Profile/sections/VolunteerProfileDocument/styles.ts
@@ -132,23 +132,6 @@ export const StatusBadge = styled.div<{ $status: "uploaded" | "missing" }>`
   background: ${(props) => (props.$status === "uploaded" ? "var(--color-green-100)" : "var(--color-red-50)")};
 `;
 
-export const ActionButton = styled.button<{ $disabled?: boolean }>`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: var(--document-section-action-button-size);
-  height: var(--document-section-action-button-size);
-  border: none;
-  background: transparent;
-  cursor: ${(props) => (props.$disabled ? "not-allowed" : "pointer")};
-  padding: 0;
-  color: ${(props) => (props.$disabled ? "var(--color-grey-200)" : "var(--color-midnight)")};
-
-  &:hover:not(:disabled) {
-    opacity: var(--document-section-action-button-hover-opacity);
-  }
-`;
-
 export const ActionCell = styled(Cell).attrs<{ $width?: string; $align?: string }>((props) => ({
   $width: props.$width ?? ACTION_COLUMN_WIDTH,
   $align: props.$align ?? "center",
